refactor: simplify TypeDefinition, Registry impl
