--- service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/organization/organizationtype/BusinessUnitOrganizationType.java
@@ -1,16 +1,0 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.organization.organizationtype;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class BusinessUnitOrganizationType implements OrganizationTypeDefinition {
-    @Override
-    public OrganizationType getOrganizationType() {
-        return new OrganizationType("BUSINESS_UNIT");
-    }
-
-    @Override
-    public String getDisplayName() {
-        return "Business Unit";
-    }
-}
