--- service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/organization/organizationtype/CompanyOrganizationType.java
@@ -1,16 +1,0 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.organization.organizationtype;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class CompanyOrganizationType implements OrganizationTypeDefinition {
-    @Override
-    public OrganizationType getOrganizationType() {
-        return new OrganizationType("COMPANY");
-    }
-
-    @Override
-    public String getDisplayName() {
-        return "Company";
-    }
-}
