--- service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/organization/organizationtype/DepartmentOrganizationType.java
@@ -1,16 +1,0 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.organization.organizationtype;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class DepartmentOrganizationType implements OrganizationTypeDefinition {
-    @Override
-    public OrganizationType getOrganizationType() {
-        return new OrganizationType("DEPARTMENT");
-    }
-
-    @Override
-    public String getDisplayName() {
-        return "Department";
-    }
-}
