--- service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/organization/organizationtype/NonProfitOrganizationType.java
@@ -1,16 +1,0 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.organization.organizationtype;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class NonProfitOrganizationType implements OrganizationTypeDefinition {
-    @Override
-    public OrganizationType getOrganizationType() {
-        return new OrganizationType("NON_PROFIT_ORGANIZATION");
-    }
-
-    @Override
-    public String getDisplayName() {
-        return "Non Profit Organization";
-    }
-}
