+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/organization/organizationtype/definition/DepartmentOrganizationType.java
@@ -1,0 +1,13 @@
+package io.commercestacksolutions.priceproviderservice.dataaccess.organization.organizationtype.definition;
+
+import io.commercestacksolutions.priceproviderservice.dataaccess.organization.organizationtype.OrganizationType;
+import io.commercestacksolutions.priceproviderservice.dataaccess.organization.organizationtype.OrganizationTypeDefinition;
+import org.springframework.stereotype.Component;
+
+@Component
+public class DepartmentOrganizationType implements OrganizationTypeDefinition {
+    @Override
+    public OrganizationType getOrganizationType() {
+        return new OrganizationType("DEPARTMENT");
+    }
+}
