--- service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/pricerow/pricetype/MaterialCostPriceType.java
@@ -1,16 +1,0 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.pricetype;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class MaterialCostPriceType implements PriceTypeDefinition {
-    @Override
-    public PriceType getPriceType() {
-        return new PriceType("MATERIAL_COST");
-    }
-
-    @Override
-    public String getDisplayName() {
-        return "Material Cost";
-    }
-}
