--- service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/pricerow/pricetype/RentalDailyRatePriceType.java
@@ -1,16 +1,0 @@
-package io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.pricetype;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class RentalDailyRatePriceType implements PriceTypeDefinition {
-    @Override
-    public PriceType getPriceType() {
-        return new PriceType("RENTAL_DAILY_RATE");
-    }
-
-    @Override
-    public String getDisplayName() {
-        return "Rental Daily Rate";
-    }
-}
