+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/dataaccess/pricerow/pricetype/definition/MaterialCostPriceType.java
@@ -1,0 +1,13 @@
+package io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.pricetype.definition;
+
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.pricetype.PriceType;
+import io.commercestacksolutions.priceproviderservice.dataaccess.pricerow.pricetype.PriceTypeDefinition;
+import org.springframework.stereotype.Component;
+
+@Component
+public class MaterialCostPriceType implements PriceTypeDefinition {
+    @Override
+    public PriceType getPriceType() {
+        return new PriceType("MATERIAL_COST");
+    }
+}
